refactor: 💡 Refactor appointment-slice to use redux for errors
