+++ package.json
@@ -5,7 +5,7 @@
   "private": false,
   "license": "MIT",
   "dependencies": {
-    "@hospitalrun/components": "^1.3.0",
+    "@hospitalrun/components": "^1.5.0",
     "@reduxjs/toolkit": "~1.3.0",
     "@types/pouchdb-find": "~6.3.4",
     "bootstrap": "~4.4.1",
