+++ src/__tests__/scheduling/appointments/Appointments.test.tsx
@@ -1,92 +1,207 @@
 import '../../../__mocks__/matchMediaMock'
 import React from 'react'
-import { mount } from 'enzyme'
-import { MemoryRouter } from 'react-router-dom'
 import { Provider } from 'react-redux'
-import Appointments from 'scheduling/appointments/Appointments'
+import { MemoryRouter } from 'react-router'
 import configureMockStore from 'redux-mock-store'
+import { mount } from 'enzyme'
 import thunk from 'redux-thunk'
-import { Calendar } from '@hospitalrun/components'
-import { act } from '@testing-library/react'
-import PatientRepository from 'clients/db/PatientRepository'
-import { mocked } from 'ts-jest/utils'
-import Patient from 'model/Patient'
-import * as ButtonBarProvider from 'page-header/ButtonBarProvider'
-import AppointmentRepository from 'clients/db/AppointmentRepository'
-import Appointment from 'model/Appointment'
-import * as titleUtil from '../../../page-header/useTitle'
+import { act } from 'react-dom/test-utils'
+import NewAppointment from 'scheduling/appointments/new/NewAppointment'
+import EditAppointment from 'scheduling/appointments/edit/EditAppointment'
+import ViewAppointments from 'scheduling/appointments/ViewAppointments'
+import Permissions from '../../../model/Permissions'
+import HospitalRun from '../../../HospitalRun'
+import { addBreadcrumbs } from '../../../breadcrumbs/breadcrumbs-slice'
+import Dashboard from '../../../dashboard/Dashboard'
+import PatientRepository from '../../../clients/db/PatientRepository'
+import AppointmentRepository from '../../../clients/db/AppointmentRepository'
+import Patient from '../../../model/Patient'
+import Appointment from '../../../model/Appointment'
 
-describe('Appointments', () => {
-  const expectedAppointments = [
-    {
-      id: '123',
-      rev: '1',
-      patientId: '1234',
-      startDateTime: new Date().toISOString(),
-      endDateTime: new Date().toISOString(),
-      location: 'location',
-      reason: 'reason',
-    },
-  ] as Appointment[]
+const mockStore = configureMockStore([thunk])
 
-  const setup = async () => {
-    jest.spyOn(AppointmentRepository, 'findAll').mockResolvedValue(expectedAppointments)
-    jest.spyOn(PatientRepository, 'find')
-    const mockedPatientRepository = mocked(PatientRepository, true)
-    mockedPatientRepository.find.mockResolvedValue({
-      id: '123',
-      fullName: 'patient full name',
-    } as Patient)
-    const mockStore = configureMockStore([thunk])
-    return mount(
-      <Provider store={mockStore({ appointments: { appointments: expectedAppointments } })}>
+describe('/appointments', () => {
+  it('should render the appointments screen when /appointments is accessed', async () => {
+    const store = mockStore({
+      title: 'test',
+      user: { permissions: [Permissions.ReadAppointments] },
+      appointments: { appointments: [] },
+      breadcrumbs: { breadcrumbs: [] },
+      components: { sidebarCollapsed: false },
+    })
+
+    const wrapper = mount(
+      <Provider store={store}>
         <MemoryRouter initialEntries={['/appointments']}>
-          <Appointments />
+          <HospitalRun />
         </MemoryRouter>
       </Provider>,
     )
-  }
 
-  it('should use "Appointments" as the header', async () => {
-    jest.spyOn(titleUtil, 'default')
     await act(async () => {
-      await setup()
+      wrapper.update()
     })
-    expect(titleUtil.default).toHaveBeenCalledWith('scheduling.appointments.label')
+
+    expect(wrapper.find(ViewAppointments)).toHaveLength(1)
+
+    expect(store.getActions()).toContainEqual(
+      addBreadcrumbs([
+        { i18nKey: 'scheduling.appointments.label', location: '/appointments' },
+        { i18nKey: 'dashboard.label', location: '/' },
+      ]),
+    )
   })
 
-  it('should add a "New Appointment" button to the button tool bar', async () => {
-    jest.spyOn(ButtonBarProvider, 'useButtonToolbarSetter')
-    const setButtonToolBarSpy = jest.fn()
-    mocked(ButtonBarProvider).useButtonToolbarSetter.mockReturnValue(setButtonToolBarSpy)
+  it('should render the Dashboard when the user does not have read appointment privileges', () => {
+    const wrapper = mount(
+      <Provider
+        store={mockStore({
+          title: 'test',
+          user: { permissions: [] },
+          breadcrumbs: { breadcrumbs: [] },
+          components: { sidebarCollapsed: false },
+        })}
+      >
+        <MemoryRouter initialEntries={['/appointments']}>
+          <HospitalRun />
+        </MemoryRouter>
+      </Provider>,
+    )
 
-    await act(async () => {
-      await setup()
-    })
-
-    const actualButtons: React.ReactNode[] = setButtonToolBarSpy.mock.calls[0][0]
-    expect((actualButtons[0] as any).props.children).toEqual('scheduling.appointments.new')
+    expect(wrapper.find(Dashboard)).toHaveLength(1)
   })
+})
 
-  it('should render a calendar with the proper events', async () => {
-    let wrapper: any
-    await act(async () => {
-      wrapper = await setup()
+describe('/appointments/new', () => {
+  it('should render the new appointment screen when /appointments/new is accessed', async () => {
+    const store = mockStore({
+      title: 'test',
+      user: { permissions: [Permissions.WriteAppointments] },
+      appointment: {},
+      breadcrumbs: { breadcrumbs: [] },
+      components: { sidebarCollapsed: false },
     })
+
+    const wrapper = mount(
+      <Provider store={store}>
+        <MemoryRouter initialEntries={['/appointments/new']}>
+          <HospitalRun />
+        </MemoryRouter>
+      </Provider>,
+    )
+
     wrapper.update()
 
-    const expectedEvents = [
-      {
-        id: expectedAppointments[0].id,
-        start: new Date(expectedAppointments[0].startDateTime),
-        end: new Date(expectedAppointments[0].endDateTime),
-        title: 'patient full name',
-        allDay: false,
-      },
-    ]
+    expect(wrapper.find(NewAppointment)).toHaveLength(1)
+    expect(store.getActions()).toContainEqual(
+      addBreadcrumbs([
+        { i18nKey: 'scheduling.appointments.label', location: '/appointments' },
+        { i18nKey: 'scheduling.appointments.new', location: '/appointments/new' },
+        { i18nKey: 'dashboard.label', location: '/' },
+      ]),
+    )
+  })
 
-    const calendar = wrapper.find(Calendar)
-    expect(calendar).toHaveLength(1)
-    expect(calendar.prop('events')).toEqual(expectedEvents)
+  it('should render the Dashboard when the user does not have read appointment privileges', () => {
+    const wrapper = mount(
+      <Provider
+        store={mockStore({
+          title: 'test',
+          user: { permissions: [] },
+          breadcrumbs: { breadcrumbs: [] },
+          components: { sidebarCollapsed: false },
+        })}
+      >
+        <MemoryRouter initialEntries={['/appointments/new']}>
+          <HospitalRun />
+        </MemoryRouter>
+      </Provider>,
+    )
+
+    expect(wrapper.find(Dashboard)).toHaveLength(1)
+  })
+})
+
+describe('/appointments/edit/:id', () => {
+  it('should render the edit appointment screen when /appointments/edit/:id is accessed', () => {
+    const appointment = {
+      id: '123',
+      patientId: '456',
+    } as Appointment
+
+    const patient = {
+      id: '456',
+    } as Patient
+
+    jest.spyOn(AppointmentRepository, 'find').mockResolvedValue(appointment)
+    jest.spyOn(PatientRepository, 'find').mockResolvedValue(patient)
+
+    const store = mockStore({
+      title: 'test',
+      user: { permissions: [Permissions.WriteAppointments, Permissions.ReadAppointments] },
+      appointment: { appointment, patient: {} as Patient },
+      breadcrumbs: { breadcrumbs: [] },
+      components: { sidebarCollapsed: false },
+    })
+
+    const wrapper = mount(
+      <Provider store={store}>
+        <MemoryRouter initialEntries={['/appointments/edit/123']}>
+          <HospitalRun />
+        </MemoryRouter>
+      </Provider>,
+    )
+
+    expect(wrapper.find(EditAppointment)).toHaveLength(1)
+
+    expect(store.getActions()).toContainEqual(
+      addBreadcrumbs([
+        { i18nKey: 'scheduling.appointments.label', location: '/appointments' },
+        { text: '123', location: '/appointments/123' },
+        {
+          i18nKey: 'scheduling.appointments.editAppointment',
+          location: '/appointments/edit/123',
+        },
+        { i18nKey: 'dashboard.label', location: '/' },
+      ]),
+    )
+  })
+
+  it('should render the Dashboard when the user does not have read appointment privileges', () => {
+    const wrapper = mount(
+      <Provider
+        store={mockStore({
+          title: 'test',
+          user: { permissions: [Permissions.WriteAppointments] },
+          breadcrumbs: { breadcrumbs: [] },
+          components: { sidebarCollapsed: false },
+        })}
+      >
+        <MemoryRouter initialEntries={['/appointments/edit/123']}>
+          <HospitalRun />
+        </MemoryRouter>
+      </Provider>,
+    )
+
+    expect(wrapper.find(Dashboard)).toHaveLength(1)
+  })
+
+  it('should render the Dashboard when the user does not have write appointment privileges', () => {
+    const wrapper = mount(
+      <Provider
+        store={mockStore({
+          title: 'test',
+          user: { permissions: [Permissions.ReadAppointments] },
+          breadcrumbs: { breadcrumbs: [] },
+          components: { sidebarCollapsed: false },
+        })}
+      >
+        <MemoryRouter initialEntries={['/appointments/edit/123']}>
+          <HospitalRun />
+        </MemoryRouter>
+      </Provider>,
+    )
+
+    expect(wrapper.find(Dashboard)).toHaveLength(1)
   })
 })
