+++ src/__tests__/scheduling/appointments/appointment-slice.test.ts
@@ -5,15 +5,19 @@ import AppointmentRepository from 'clients/db/AppointmentRepository'
 import { mocked } from 'ts-jest/utils'
 import PatientRepository from 'clients/db/PatientRepository'
 import Patient from 'model/Patient'
+import { subDays } from 'date-fns'
+
 import appointment, {
   fetchAppointmentStart,
   fetchAppointmentSuccess,
   fetchAppointment,
   createAppointmentStart,
   createAppointmentSuccess,
+  createAppointmentError,
   createAppointment,
   updateAppointmentStart,
   updateAppointmentSuccess,
+  updateAppointmentError,
   updateAppointment,
   deleteAppointment,
   deleteAppointmentStart,
@@ -26,14 +30,14 @@ describe('appointment slice', () => {
       const appointmentStore = appointment(undefined, {} as AnyAction)
 
       expect(appointmentStore.appointment).toEqual({} as Appointment)
-      expect(appointmentStore.isLoading).toBeFalsy()
+      expect(appointmentStore.status).toEqual('loading')
     })
     it('should handle the CREATE_APPOINTMENT_START action', () => {
       const appointmentStore = appointment(undefined, {
         type: createAppointmentStart.type,
       })
 
-      expect(appointmentStore.isLoading).toBeTruthy()
+      expect(appointmentStore.status).toEqual('loading')
     })
 
     it('should handle the CREATE_APPOINTMENT_SUCCESS action', () => {
@@ -41,7 +45,7 @@ describe('appointment slice', () => {
         type: createAppointmentSuccess.type,
       })
 
-      expect(appointmentStore.isLoading).toBeFalsy()
+      expect(appointmentStore.status).toEqual('completed')
     })
 
     it('should handle the UPDATE_APPOINTMENT_START action', () => {
@@ -49,7 +53,7 @@ describe('appointment slice', () => {
         type: updateAppointmentStart.type,
       })
 
-      expect(appointmentStore.isLoading).toBeTruthy()
+      expect(appointmentStore.status).toEqual('loading')
     })
 
     it('should handle the UPDATE_APPOINTMENT_SUCCESS action', () => {
@@ -68,7 +72,7 @@ describe('appointment slice', () => {
         },
       })
 
-      expect(appointmentStore.isLoading).toBeFalsy()
+      expect(appointmentStore.status).toEqual('completed')
       expect(appointmentStore.appointment).toEqual(expectedAppointment)
     })
 
@@ -77,7 +81,7 @@ describe('appointment slice', () => {
         type: fetchAppointmentStart.type,
       })
 
-      expect(appointmentStore.isLoading).toBeTruthy()
+      expect(appointmentStore.status).toEqual('loading')
     })
     it('should handle the FETCH_APPOINTMENT_SUCCESS action', () => {
       const expectedAppointment = {
@@ -96,7 +100,7 @@ describe('appointment slice', () => {
         payload: { appointment: expectedAppointment, patient: expectedPatient },
       })
 
-      expect(appointmentStore.isLoading).toBeFalsy()
+      expect(appointmentStore.status).toEqual('completed')
       expect(appointmentStore.appointment).toEqual(expectedAppointment)
       expect(appointmentStore.patient).toEqual(expectedPatient)
     })
@@ -106,7 +110,7 @@ describe('appointment slice', () => {
         type: deleteAppointmentStart.type,
       })
 
-      expect(appointmentStore.isLoading).toBeTruthy()
+      expect(appointmentStore.status).toEqual('loading')
     })
 
     it('should handle the DELETE_APPOINTMENT_SUCCESS action', () => {
@@ -114,7 +118,7 @@ describe('appointment slice', () => {
         type: deleteAppointmentSuccess.type,
       })
 
-      expect(appointmentStore.isLoading).toBeFalsy()
+      expect(appointmentStore.status).toEqual('completed')
     })
   })
 
@@ -181,6 +185,29 @@ describe('appointment slice', () => {
       await createAppointment(expectedAppointment, onSuccessSpy)(dispatch, getState, null)
 
       expect(onSuccessSpy).toHaveBeenCalledWith(expectedSavedAppointment)
+    })
+
+    it('should validate the appointment', async () => {
+      const dispatch = jest.fn()
+      const getState = jest.fn()
+      const expectedAppointment = {
+        startDateTime: new Date().toISOString(),
+        endDateTime: subDays(new Date(), 5).toISOString(),
+        location: 'location',
+        type: 'type',
+        reason: 'reason',
+      } as Appointment
+
+      await createAppointment(expectedAppointment)(dispatch, getState, null)
+
+      expect(dispatch).toHaveBeenNthCalledWith(2, {
+        type: createAppointmentError.type,
+        payload: {
+          message: 'scheduling.appointment.errors.createAppointmentError',
+          patient: 'scheduling.appointment.errors.patientRequired',
+          startDateTime: 'scheduling.appointment.errors.startDateMustBeBeforeEndDate',
+        },
+      })
     })
   })
 
@@ -303,8 +330,16 @@ describe('appointment slice', () => {
       const dispatch = jest.fn()
       const getState = jest.fn()
       jest.spyOn(AppointmentRepository, 'saveOrUpdate')
-      const expectedAppointmentId = 'sliceId9'
-      const expectedAppointment = { id: expectedAppointmentId } as Appointment
+
+      const expectedAppointment = {
+        patientId: 'sliceId9',
+        startDateTime: new Date().toISOString(),
+        endDateTime: new Date().toISOString(),
+        location: 'location',
+        type: 'type',
+        reason: 'reason',
+      } as Appointment
+
       const mockedAppointmentRepository = mocked(AppointmentRepository, true)
       mockedAppointmentRepository.saveOrUpdate.mockResolvedValue(expectedAppointment)
 
@@ -317,8 +352,16 @@ describe('appointment slice', () => {
       const dispatch = jest.fn()
       const getState = jest.fn()
       jest.spyOn(AppointmentRepository, 'saveOrUpdate')
-      const expectedAppointmentId = 'sliceId10'
-      const expectedAppointment = { id: expectedAppointmentId } as Appointment
+
+      const expectedAppointment = {
+        patientId: 'sliceId10',
+        startDateTime: new Date().toISOString(),
+        endDateTime: new Date().toISOString(),
+        location: 'location',
+        type: 'type',
+        reason: 'reason',
+      } as Appointment
+
       const mockedAppointmentRepository = mocked(AppointmentRepository, true)
       mockedAppointmentRepository.saveOrUpdate.mockResolvedValue(expectedAppointment)
 
@@ -331,8 +374,16 @@ describe('appointment slice', () => {
       const dispatch = jest.fn()
       const getState = jest.fn()
       jest.spyOn(AppointmentRepository, 'saveOrUpdate')
-      const expectedAppointmentId = 'sliceId11'
-      const expectedAppointment = { id: expectedAppointmentId } as Appointment
+
+      const expectedAppointment = {
+        patientId: 'sliceId11',
+        startDateTime: new Date().toISOString(),
+        endDateTime: new Date().toISOString(),
+        location: 'location',
+        type: 'type',
+        reason: 'reason',
+      } as Appointment
+
       const mockedAppointmentRepository = mocked(AppointmentRepository, true)
       mockedAppointmentRepository.saveOrUpdate.mockResolvedValue(expectedAppointment)
 
@@ -349,14 +400,45 @@ describe('appointment slice', () => {
       const dispatch = jest.fn()
       const getState = jest.fn()
       jest.spyOn(AppointmentRepository, 'saveOrUpdate')
-      const expectedAppointmentId = 'sliceId11'
-      const expectedAppointment = { id: expectedAppointmentId } as Appointment
+
+      const expectedAppointment = {
+        patientId: 'sliceId12',
+        startDateTime: new Date().toISOString(),
+        endDateTime: new Date().toISOString(),
+        location: 'location',
+        type: 'type',
+        reason: 'reason',
+      } as Appointment
+
       const mockedAppointmentRepository = mocked(AppointmentRepository, true)
       mockedAppointmentRepository.saveOrUpdate.mockResolvedValue(expectedAppointment)
 
       await updateAppointment(expectedAppointment, onSuccessSpy)(dispatch, getState, null)
 
       expect(onSuccessSpy).toHaveBeenCalledWith(expectedAppointment)
+    })
+
+    it('should validate the appointment', async () => {
+      const dispatch = jest.fn()
+      const getState = jest.fn()
+      const expectedAppointment = {
+        startDateTime: new Date().toISOString(),
+        endDateTime: subDays(new Date(), 5).toISOString(),
+        location: 'location',
+        type: 'type',
+        reason: 'reason',
+      } as Appointment
+
+      await updateAppointment(expectedAppointment)(dispatch, getState, null)
+
+      expect(dispatch).toHaveBeenNthCalledWith(2, {
+        type: updateAppointmentError.type,
+        payload: {
+          message: 'scheduling.appointment.errors.updateAppointmentError',
+          patient: 'scheduling.appointment.errors.patientRequired',
+          startDateTime: 'scheduling.appointment.errors.startDateMustBeBeforeEndDate',
+        },
+      })
     })
   })
 })
