+++ src/locales/enUs/translations/scheduling/index.ts
@@ -22,14 +22,16 @@ export default {
         walkIn: 'Walk In',
       },
       errors: {
+        createAppointmentError: 'Could not create new appointment.',
+        updateAppointmentError: 'Could not update appointment.',
         patientRequired: 'Patient is required.',
-        errorCreatingAppointment: 'Error Creating Appointment!',
         startDateMustBeBeforeEndDate: 'Start Time must be before End Time.',
       },
       reason: 'Reason',
       patient: 'Patient',
       deleteConfirmationMessage: 'Are you sure that you want to delete this appointment?',
       successfullyCreated: 'Successfully created appointment.',
+      successfullyDeleted: 'Successfully deleted appointment.',
     },
   },
 }
