+++ src/scheduling/appointments/appointment-slice.ts
@@ -4,48 +4,82 @@ import { AppThunk } from 'store'
 import AppointmentRepository from 'clients/db/AppointmentRepository'
 import Patient from 'model/Patient'
 import PatientRepository from 'clients/db/PatientRepository'
+import { isBefore } from 'date-fns'
+import _ from 'lodash'
+
+function validateAppointment(appointment: Appointment) {
+  const err: Error = {}
+
+  if (!appointment.patientId) {
+    err.patient = 'scheduling.appointment.errors.patientRequired'
+  }
+
+  if (isBefore(new Date(appointment.endDateTime), new Date(appointment.startDateTime))) {
+    err.startDateTime = 'scheduling.appointment.errors.startDateMustBeBeforeEndDate'
+  }
+
+  return err
+}
+
+interface Error {
+  patient?: string
+  startDateTime?: string
+  message?: string
+}
 
 interface AppointmentState {
+  error: Error
   appointment: Appointment
   patient: Patient
-  isLoading: boolean
+  status: 'loading' | 'error' | 'completed'
 }
 
-const initialAppointmentState = {
+const initialState: AppointmentState = {
+  error: {},
   appointment: {} as Appointment,
   patient: {} as Patient,
-  isLoading: false,
+  status: 'loading',
 }
 
-function startLoading(state: AppointmentState) {
-  state.isLoading = true
+function start(state: AppointmentState) {
+  state.status = 'loading'
+}
+
+function finish(state: AppointmentState, { payload }: PayloadAction<Appointment>) {
+  state.status = 'completed'
+  state.appointment = payload
+  state.error = {}
+}
+
+function error(state: AppointmentState, { payload }: PayloadAction<Error>) {
+  state.status = 'error'
+  state.error = payload
 }
 
 const appointmentSlice = createSlice({
   name: 'appointment',
-  initialState: initialAppointmentState,
+  initialState,
   reducers: {
-    fetchAppointmentStart: startLoading,
-    createAppointmentStart: startLoading,
-    updateAppointmentStart: startLoading,
-    deleteAppointmentStart: startLoading,
-    deleteAppointmentSuccess: (state: AppointmentState) => {
-      state.isLoading = false
-    },
+    fetchAppointmentStart: start,
     fetchAppointmentSuccess: (
-      state,
+      state: AppointmentState,
       { payload }: PayloadAction<{ appointment: Appointment; patient: Patient }>,
     ) => {
-      state.isLoading = false
+      state.status = 'completed'
       state.appointment = payload.appointment
       state.patient = payload.patient
     },
+    createAppointmentStart: start,
     createAppointmentSuccess(state) {
-      state.isLoading = false
+      state.status = 'completed'
     },
-    updateAppointmentSuccess(state, { payload }: PayloadAction<Appointment>) {
-      state.isLoading = false
-      state.appointment = payload
+    createAppointmentError: error,
+    updateAppointmentStart: start,
+    updateAppointmentSuccess: finish,
+    updateAppointmentError: error,
+    deleteAppointmentStart: start,
+    deleteAppointmentSuccess(state) {
+      state.status = 'completed'
     },
   },
 })
@@ -53,8 +87,10 @@ const appointmentSlice = createSlice({
 export const {
   createAppointmentStart,
   createAppointmentSuccess,
+  createAppointmentError,
   updateAppointmentStart,
   updateAppointmentSuccess,
+  updateAppointmentError,
   fetchAppointmentStart,
   fetchAppointmentSuccess,
   deleteAppointmentStart,
@@ -74,10 +110,17 @@ export const createAppointment = (
   onSuccess?: (appointment: Appointment) => void,
 ): AppThunk => async (dispatch) => {
   dispatch(createAppointmentStart())
-  const newAppointment = await AppointmentRepository.save(appointment)
-  dispatch(createAppointmentSuccess())
-  if (onSuccess) {
-    onSuccess(newAppointment)
+  const newAppointmentError = validateAppointment(appointment)
+
+  if (_.isEmpty(newAppointmentError)) {
+    const newAppointment = await AppointmentRepository.save(appointment)
+    dispatch(createAppointmentSuccess())
+    if (onSuccess) {
+      onSuccess(newAppointment)
+    }
+  } else {
+    newAppointmentError.message = 'scheduling.appointment.errors.createAppointmentError'
+    dispatch(createAppointmentError(newAppointmentError))
   }
 }
 
@@ -86,11 +129,18 @@ export const updateAppointment = (
   onSuccess?: (appointment: Appointment) => void,
 ): AppThunk => async (dispatch) => {
   dispatch(updateAppointmentStart())
-  const updatedAppointment = await AppointmentRepository.saveOrUpdate(appointment)
-  dispatch(updateAppointmentSuccess(updatedAppointment))
+  const updatedAppointmentError = validateAppointment(appointment)
 
-  if (onSuccess) {
-    onSuccess(updatedAppointment)
+  if (_.isEmpty(updatedAppointmentError)) {
+    const updatedAppointment = await AppointmentRepository.saveOrUpdate(appointment)
+    dispatch(updateAppointmentSuccess(updatedAppointment))
+
+    if (onSuccess) {
+      onSuccess(updatedAppointment)
+    }
+  } else {
+    updatedAppointmentError.message = 'scheduling.appointment.errors.updateAppointmentError'
+    dispatch(updateAppointmentError(updatedAppointmentError))
   }
 }
